make things work using resolve.extensions
This adds .scss to the resolve.extensions array, which makes includes like `foo`  try to resolve to `foo.scss` after trying `foo.js` and `foo.jsx`.

This has issues because style files cannot share a name with js or jsx files (or anything else higher in the chain). Try adding a style.js file to this checkout and see what happens.